% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppXsimd.R
\name{getAVX512Flags}
\alias{getAVX512Flags}
\title{Concatenate supported AVX512 compiler flags for system CPU}
\usage{
getAVX512Flags()
}
\value{
String for compiler flags
}
\description{
Concatenate supported AVX512 compiler flags for system CPU
}
