% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppXsimd.R
\name{getNEONFlags}
\alias{getNEONFlags}
\title{Concatenate supported NEON compiler flags for system CPU}
\usage{
getNEONFlags()
}
\value{
String for compiler flags
}
\description{
Concatenate supported NEON compiler flags for system CPU
}
