% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSimdFeatures}
\alias{getSimdFeatures}
\title{Poll OS and CPU for SIMD feature support}
\usage{
getSimdFeatures()
}
\value{
List of operating sytem (OS) and hardware (HW) feature support; 
see CPUID Wiki page for flag definitions
}
\description{
Execute CPUID to poll operating system and central processing unit for 
single instruction, multiple data feature support.
}
\references{
\url{https://en.wikipedia.org/wiki/CPUID}
}
