% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppXsimd.R
\name{getSSEFlags}
\alias{getSSEFlags}
\title{Concatenation supported SSE compiler flags for system CPU}
\usage{
getSSEFlags()
}
\value{
String for compiler flags
}
\description{
Concatenation supported SSE compiler flags for system CPU
}
