% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppXsimd.R
\name{supportsAVX512}
\alias{supportsAVX512}
\title{Determine if CPU supports AVX512 SIMD instructions}
\usage{
supportsAVX512()
}
\value{
Boolean
}
\description{
Determine if CPU supports AVX512 SIMD instructions
}
