% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppXsimd.R
\name{supportsSSE42}
\alias{supportsSSE42}
\title{Determine if CPU supports SSE42 SIMD instructions}
\usage{
supportsSSE42()
}
\value{
Boolean
}
\description{
Determine if CPU supports SSE42 SIMD instructions
}
