% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppXsimd.R
\name{getAVXFlags}
\alias{getAVXFlags}
\title{Concatenate supported AVX compiler flags for system CPU}
\usage{
getAVXFlags()
}
\value{
String for compiler flags
}
\description{
Concatenate supported AVX compiler flags for system CPU
}
