% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppXsimd.R
\docType{package}
\name{RcppXsimd}
\alias{RcppXsimd}
\alias{RcppXsimd-package}
\title{RcppXsimd: Rcpp wrapper to Xsimd}
\description{
The RcppXsimd package wrappers the header-only C++ Xsimd library that 
provides parallelized math implementations using SIMD
}
